server: drop sequelize
