--- packages/server/.sequelizerc
@@ -1,8 +1,0 @@
-const path = require('path')
-
-module.exports = {
-  "config": path.resolve('./src/database/', 'config.js'),
-  "models-path": path.resolve('./src/models'),
-  "seeders-path": path.resolve('./src/database/seeders'),
-  "migrations-path": path.resolve('./src/database/migrations')
-}
