+++ packages/server/src/index.js
@@ -12,14 +12,12 @@ import { createMiddleware, signalIsUp } from '@promster/express'
 import routes from './routes'
 import errorHandler from './errors/error-handler'
 import corsMiddleware from './helpers/cors-middleware'
-import checkConnection from './database/check-connection'
 import sessionMiddleware from './helpers/session-middleware'
 import notFoundMiddleware from './helpers/not-found-middleware'
 import sanitizeAddressMiddleware from './helpers/sanitize-address-middleware'
 
 // Load env variables and check DB connection
 dotenv.config()
-checkConnection()
 
 // Set up express layers
 const app = express()
