+++ packages/server/src/models/objection/UserEmail.js
@@ -1,8 +1,10 @@
 import BaseModel from './BaseModel'
+
 export default class UserEmail extends BaseModel {
   static get tableName() {
     return 'UserEmails'
   }
+
   static get relationMappings() {
     return {
       users: {
