+++ packages/server/src/models/objection/UserEmailVerificationToken.js
@@ -1,8 +1,10 @@
 import BaseModel from './BaseModel'
+
 export default class UserEmailVerificationToken extends BaseModel {
   static get tableName() {
     return 'UserEmailVerificationTokens'
   }
+
   static get relationMappings() {
     return {
       user: {
